benchlab/wigidash/requirements.txt
Pillow=>=11.3.0
matplotlib>=3.10.7
numpy=2.3.3
pyusb>=1.3.1
pyserial>=3.5